@@ -15,42 +15,44 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.3+9
+version: 1.0.4+10
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.8.4 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  built_value: ^7.1.0
   http: ^0.12.1
   random_string: ^2.1.0
   expandable_bottom_sheet: ^0.2.1+1
   expandable: ^4.1.4
   path_provider: ^1.6.11
-  hive: ^1.4.1+1
-  hive_flutter: ^0.3.0+2
-  built_collection: ^4.3.2
-  flutter_bloc: ^5.0.1
+  hive: ^1.4.4+1
+  hive_flutter: ^0.3.1
+  flutter_bloc: ^6.1.2
+  bloc: ^6.1.1
+  google_sign_in: ^4.5.6
   kiwi: ^2.0.0
   flutter_native_admob: ^2.1.0
-  firebase_messaging: ^6.0.16
   eva_icons_flutter: ^2.0.0
   url_launcher: ^5.5.0
   flutter_email_sender: ^3.0.1
-
+  firebase_core: ^0.5.1
+  firebase_messaging: ^6.0.16
+  firebase_auth: ^0.18.2
+  firebase_database: ^4.2.0
+  json_serializable: ^3.5.0
+  flutter_auth_buttons: ^0.10.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.3
-dependency_overrides:
-  dartx: ^0.3.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  hive_generator: ^0.7.0+2
-  build_runner: ^1.10.0
-  built_value_generator: ^7.1.0
+  hive_generator: ^0.8.2
+  build_runner: ^1.11.1
   flutter_launcher_icons: "^0.7.3"
   device_preview: ^0.4.7
 
