@@ -3,9 +3,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_email_sender/flutter_email_sender.dart';
 import 'package:hive/hive.dart';
-import 'package:random_waifu/ui/bloc/InformationBloc/InformationBloc.dart';
+import 'package:random_waifu/ui/bloc/InformationBloc/InformationCubit.dart';
+import 'package:random_waifu/ui/pages/AccountPage.dart';
 import 'package:random_waifu/ui/widgets/CharacterInformation/CharacterImage.dart';
 import 'package:random_waifu/ui/widgets/CharacterInformation/CharacterNameWidget.dart';
+import 'package:random_waifu/ui/widgets/shared/ConfirmDialogWidget.dart';
 
 class CharacterInformationWidget extends StatelessWidget {
   final String imageUrl;
@@ -29,64 +31,68 @@ class CharacterInformationWidget extends StatelessWidget {
           imageUrl: imageUrl,
         ),
         Container(
-          margin: EdgeInsets.only(top:30,bottom: 30),
-          alignment: Alignment.topLeft,
-          width: double.infinity,
-          child: IconButton(
-            icon: Icon(Icons.feedback),
-            color: Colors.white,
-            onPressed: () => {
-              showDialog(
-                context: context,
-                builder: (BuildContext context) {
-                  return AlertDialog(
-                    title: Text("Not a waifu report"),
-                    content: Text(
-                        "Would you like to report this character?"),
-                    actions: <Widget>[
-                      RaisedButton(
-                        color: Colors.indigoAccent,
-                        onPressed: () async {
-                          final Email email = Email(
-                            body:
-                                'Character name: ${this.name}\nCharacter id: ${this.malId}',
-                            subject: 'Not a waifu report :(',
-                            cc: ['logicappsfeedback@gmail.com'],
-                            isHTML: false,
-                          );
-
-                          await FlutterEmailSender.send(email);
-                          var characters = Hive.box("savedCharacters");
-                          var lastDate = Hive.box("today");
-                          await Hive.box('today').clear();
-
-                          await characters.deleteAt(characters.length - 1);
-
-                          print(characters.length);
-                          Navigator.of(context).pop();
-                          BlocProvider.of<InformationBloc>(parentContext)
-                              .getRandomWaifu();
+            margin: EdgeInsets.only(top: 30, bottom: 30),
+            alignment: Alignment.topLeft,
+            width: double.infinity,
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                IconButton(
+                  icon: Icon(Icons.feedback),
+                  color: Colors.white,
+                  onPressed: () => {
+                    ConfirmDialog().show(
+                      context: parentContext,
+                      title: "Not a waifu report",
+                      confirm: "Report",
+                      text: "Would you like to report this character?",
+                      callback: () => {
+                        reportWaifu(context),
+                      },
+                    ),
+                  },
+                ),
+                IconButton(
+                  icon: Icon(Icons.cloud),
+                  color: Colors.white,
+                  onPressed: () => {
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                        builder: (_) {
+                          return AccountPage();
                         },
-                        child: Text("Report"),
                       ),
-                      FlatButton(
-                        onPressed: () {
-                          Navigator.of(context).pop();
-                        },
-                        child: Text("Close"),
-                      ),
-                    ],
-                  );
-                },
-              ),
-            },
-          ),
-        ),
+                    )
+                  },
+                ),
+              ],
+            )),
         CharacterNameWidget(
           name: name,
           malId: malId,
         ),
       ],
     );
+  }
+
+  void reportWaifu(context) async {
+    final Email email = Email(
+      body: 'Character name: ${name}\nCharacter id: ${malId}',
+      subject: 'Not a waifu report :(',
+      cc: ['logicappsfeedback@gmail.com'],
+      isHTML: false,
+    );
+
+    await FlutterEmailSender.send(email);
+    var characters = Hive.box("savedCharacters");
+    var lastDate = Hive.box("today");
+    await Hive.box('today').clear();
+
+    await characters.deleteAt(characters.length - 1);
+
+    print(characters.length);
+    Navigator.of(context).pop();
+    BlocProvider.of<InformationCubit>(context).getRandomWaifu();
   }
 }
